backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt>=3.1.0,<4.0
python-multipart==0.0.9
pydantic[email]==2.6.1
pydantic-settings==2.1.0
httpx==0.27.0
beautifulsoup4==4.12.3
websockets==12.0
python-slugify==8.0.4
slowapi==0.1.9
